Move AgentFeedbackEditorWidget into its own module
The agent feedback widget unit test imported
`agentFeedbackEditorWidgetContribution.ts`, whose module scope calls
`registerEditorContribution(...)`. That registered the contribution for
every editor created anywhere in the unit test run, so an unrelated test
without `IAgentFeedbackService` eventually instantiated it, threw, and
leaked an undisposed `AgentFeedbackEditorWidgetContribution`.

Split the widget into a side-effect free `agentFeedbackEditorWidget.ts`
so tests can import it without registering the editor contribution.
